requirements.txt
django
datetime
dnspython
python-dotenv
gunicorn
django-tinymce
django-jsonify
django-hCaptcha
psycopg2
whitenoise
stripe
dj-database-url
